This commit changes the handling of the C include paths.
The C library include path methods originally used come from:

https://metacpan.org/pod/Module::Install::XSUtil

This component is not a part of the core Module::Install distribution
and it is not included in this distributions inc/ directory.

This commit introduces the same methods as used in Crypt::OpenSSL::X509.
Which are included in core Module::Install via Module::Install::Makefile